Add files via upload
To begin the process I created a table using MySQL, I started by naming the table finishing the line with open parenthesis, then defining the columns and mentioning the data type of the columns by text and interger. The column I named the primary key was book id used as an identifier. 

I inserted values into the ‘BOOK’ table by using the correct syntax for each row created. 
I then created a ‘SELECT ALL/SELECT * FROM BOOK’, which is use to display all the data in the Table when you run the database.